Added ability to edit todo items
The ability to edit existing todo item by double-clicking
on the required todo. Writing to the local state takes place,
then validation takes place, and if everything is successful,
a request for Hasura is made and the updated element is redrawn.

===
I rewrote todo item from <ul> => <li> format to list format input.
Created a new Loader that I use in this component when updating data.