Displaying the names of opponents in the game
After creating a game, the user is sent to the game room,
and his name is displayed in the list of players in this room.

After this, the counter changes its value and shows that one place in this game room is occupied.
After that, the user who created this game will no longer be able to enter this room.

Another user can enter the room by clicking the join button.
After that, the counter will show that both places are occupied,
and the name of the second player will be displayed in the corresponding place.