Displaying buttons in the game room
Displaying buttons in the game room depending on the user.
If the user has created a game, then he cannot leave the game room,
the button to close the game room is available to him.

The user who joined the game room, he cannot close it. He can only leave of this room.