UI development navigation field
After downloading the data about the players, each of the players
must confirm their readiness to start the game.
After pressing the Ready button, the indicator next to the name of the player
who pressed the button changes its color to green, and <Loader /> disappears.
This indicates that the player is already ready to play and is waiting
for his opponent's readiness.