Fixed wrong order of symbols
Why:
When the user made his move, the symbol was displayed.
When another user made a move, the same symbol was displayed.
After a few moves, the symbol changed for several more moves.

Solution:
When the user entered the game room, the data was updated through
the subscription, then useEffect was triggered. Which called the function
that processes the data for the mutation. But on the first download,
useEffect created unnecessary data overwrites. To solve this problem,
two checks have been created that change two values in the state object in turn.

It is impossible not to use useEffect, without it the component of
the current user will not know what another user has changed.