Fixed a bug with setting the desired symbol
Why:
The received data was processed incorrectly.
Therefore, the symbols were displayed in the wrong order.

Solution:
Changed default values on Hasura, from number to null.
This value remains until the first click. In this case, the first  symbol
is recorded in Hasur 1, which corresponds to the first symbol in the game "X".